requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.5
schedule==1.2.0
python-dotenv==1.0.0
APScheduler==3.10.4